pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agenteval"
version = "0.1.0"
description = "Evaluation framework for multi-turn conversational AI agents"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "typer>=0.9",
    "pyyaml>=6.0",
    "rich>=13.0",
    "aiosqlite>=0.19",
    "jinja2>=3.1",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.39"]
langgraph = ["langgraph>=0.2"]
openai = ["openai-agents>=0.1"]
all = ["agenteval[anthropic,langgraph,openai]"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "coverage>=7.0"]

[project.scripts]
agenteval = "agenteval.cli:app"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
